==> 02-mgm/32-determine-labels.ipynb <==
# Determine labels for geospatial entities

We have a set of rules that define the labels for our geospatial entities.
This file goes throught that list of rules to determine all of the possible labels.
The it writes the list to a file along with a numberical ID.



## Processing Setup

# Google colab setup
# import os
# from google.colab import drive
# drive.mount('/content/drive')
# project_home = '/content/drive/MyDrive/Projects/verge'
# os.chdir(project_home)

# # Local processing setup
project_home = '..'

!pwd

## Notebook setup

import pandas as pd
import numpy as np

import sys
sys.path.append(project_home)
from utils.verge import rules


## Parameters

# The name of the ROI to use.
roi_name = 'ne-laptop'

# The name of the general-purpose data directory.
data_home = '%s/data' % (project_home)

# The name of the ROI-specific data directory.
roi_home = '%s/data/%s' % (project_home, roi_name)


labels = set()
for r in rules:
    label = '%s : %s' % (r['gent_category'], r['gent_label'])
    labels.add(label)

# This is a special case, that is created differently so it is not in the rule set.
labels.add('waterway : land')

# # These will be special cases.
# labels.add('token : mask')
# labels.add('token : pad')

records = []
for k, label in enumerate(sorted(labels)):
    records.append({'id': k, 'label': label, })
    print(k, label)


fname = '%s/labels.csv' % data_home
pd.DataFrame(records).to_csv(fname, index=False)
print('wrote %s' % fname)
