==> 02-mgm/35.0-define-splits.ipynb <==
### VERGE: Vector-mode Regional Geospatial Encoding
# Split definition

Divide the input instances into training and vlidation sets.


## Processing Setup

# Google colab setup
import os
from google.colab import drive
drive.mount('/content/drive')
project_home = '/content/drive/MyDrive/Projects/verge'
os.chdir(project_home)
!pip install geo_encodings

# Local processing setup
# project_home = '..'

## Notebook setup

import pandas as pd
import numpy as np
import glob
import pickle
import os


## Parameters

# The name of the ROI to use.
roi_name = 'newengland'

# The name of the general-purpose data directory.
data_home = '%s/data' % (project_home)

# The name of the ROI-specific data directory.
roi_home = '%s/data/%s' % (project_home, roi_name)

# A unique identifier for this run. This will be a component of any
# output file names.
run_id = '201'

# Fraction of cases to use for training.
train_fraction = 0.8

# Always...
np.random.seed(5)

## Processing


fname = '%s/tiles.csv' % roi_home
tile_info = pd.read_csv(fname)
print('%d tiles' % len(tile_info))

aoi_tags = np.unique(tile_info['aoi_tag'].values)
print('%d unique AOIs' % len(aoi_tags))

# Add a column for the split, with a defult vlue of "train".
# It will be re-set below.
tile_info['split'] = 'train'
splits = tile_info['split'].values

tile_info.head()

# Replace split designation for a fraction of the AOIs.
for aoi_tag in aoi_tags:
    if np.random.random() > train_fraction:
        iok = tile_info['aoi_tag'] == aoi_tag
        splits[iok] = 'val'


tile_info['split'] = splits
tile_info[['aoi_tag', 'split']].value_counts()

tile_info['split'].value_counts()

# Save the split records
fname = '%s/splits-%s.csv' % (roi_home, run_id)
tile_info[['aoi_tag', 'tile_tag', 'split']].to_csv(fname, index=False)
print('saved splits to %s' % fname)
