==> 02-mgm/91-sandbox-npz.ipynb <==
# Sandbox for exploring use of H5 files

import numpy as np
import os

dname = 'testdata'
os.makedirs(dname, exist_ok=True)
n = 5

for i in range(n):
    row_count = 20 + np.random.randint(200)
    col_count = 422
    features = np.random.random((row_count, col_count))
    labels = [np.random.randint(10) for _ in range(row_count)]
    fname = '%s/instance-%d.npz' % (dname, i)
    np.savez_compressed(fname, features=features, labels=labels)
    print(fname)

import glob
globstring = 'testdata/*.npz'
fnames = glob.glob(globstring)
for fname in fnames:
    data = np.load(fname)
    features = data['features']
    labels = data['labels']
    print('features', features.shape, 'labels', labels.shape)

labels
